clojure/matching-brackets: Pass the first three tests
The idea for this solution is to push each opening bracket on to a stack
and pop from the stack for each closing brace. The first three tests
only deal with square brackets so checking that the opening and closing
brackets are the same time is not necessary to pass these tests.