wasm/resistor-color: Implement colorCode
Implement colorCode by maintaining an "array" of bytes representing the
offsets into the list of color names where each name starts. That allows
the input to be compared to each color name. Since the colors are in
order of their associated number, the index into the array of offsets
gives the color code.